- Fixed: remoting exception on .Net Framework (not Mono) after strong naming due to serialization restrictions. - Safemode now not disabled until all blocks are reported. - Added test for safemode - Added test for cluster restart.
git-svn-id: file:///mnt/c/Users/svgroot/Downloads/svn/repos/Jumbo/trunk@77 2f0f0f96-4b5b-11dd-99f4-89bc9257463c

+++ Tkl.Jumbo.Dfs.Test/TestDfsCluster.cs
@@ -13,14 +13,17 @@ namespace Tkl.Jumbo.Dfs.Test
     class TestDfsCluster
     {
         private AppDomain _clusterDomain;
-        private NameServerRunner _nameServerRunner;
+        private ClusterRunner _clusterRunner;
 
         public const int NameServerPort = 10000;
         public const int FirstDataServerPort = 10001;
 
-        private class NameServerRunner : MarshalByRefObject
+        private class ClusterRunner : MarshalByRefObject
         {
-            public void Run(string editLogPath, int replicationFactor, int dataNodes, int? blockSize)
+            private int _nextDataServerPort = FirstDataServerPort;
+            private string _path;
+
+            public void Run(string editLogPath, int replicationFactor, int dataServers, int? blockSize)
             {
                 //log4net.Config.BasicConfigurator.Configure(new log4net.Appender.FileAppender(new log4net.Layout.PatternLayout("%date [%thread] %-5level %logger [%property{ClientHostName}] - %message%newline"), "/home2/sgroot/jumbo/test.log") { Threshold = log4net.Core.Level.All });
                 //log4net.Config.BasicConfigurator.Configure(new log4net.Appender.FileAppender() { Layout = new log4net.Layout.PatternLayout("%date [%thread] %-5level %logger [%property{ClientHostName}] - %message%newline"), File = System.IO.Path.Combine(editLogPath, "logfile.txt"), Threshold = log4net.Core.Level.All });
@@ -35,15 +38,19 @@ namespace Tkl.Jumbo.Dfs.Test
                 if( Environment.OSVersion.Platform == PlatformID.Unix )
                     config.NameServer.ListenIPv4AndIPv6 = false;
                 NameServer.Run(config);
-                if( dataNodes > 0 )
+                _path = editLogPath;
+                StartDataServers(dataServers);
+            }
+
+            public void StartDataServers(int dataServers)
+            {
+                if( dataServers > 0 )
                 {
-                    DataServerRunner dataServerRunner = new DataServerRunner();
-                    int port = FirstDataServerPort;
-                    for( int x = 0; x < dataNodes; ++x, ++port )
+                    for( int x = 0; x < dataServers; ++x, ++_nextDataServerPort )
                     {
-                        string blocksPath = System.IO.Path.Combine(editLogPath, "blocks" + x.ToString(System.Globalization.CultureInfo.InvariantCulture));
+                        string blocksPath = System.IO.Path.Combine(_path, "blocks" + x.ToString(System.Globalization.CultureInfo.InvariantCulture));
                         System.IO.Directory.CreateDirectory(blocksPath);
-                        dataServerRunner.Run(blocksPath, port);
+                        RunDataServer(blocksPath, _nextDataServerPort);
                     }
                 }
             }
@@ -52,11 +59,8 @@ namespace Tkl.Jumbo.Dfs.Test
             {
                 NameServer.Shutdown();
             }
-        }
 
-        private class DataServerRunner : MarshalByRefObject
-        {
-            public void Run(string path, int port)
+            private void RunDataServer(string path, int port)
             {
                 DfsConfiguration config = new DfsConfiguration();
                 config.NameServer.HostName = "localhost";
@@ -65,13 +69,13 @@ namespace Tkl.Jumbo.Dfs.Test
                 config.DataServer.BlockStoragePath = path;
                 if( Environment.OSVersion.Platform == PlatformID.Unix )
                     config.DataServer.ListenIPv4AndIPv6 = false;
-                Thread t = new Thread(RunThread);
+                Thread t = new Thread(RunDataServerThread);
                 t.Name = "DataServer";
                 t.IsBackground = true;
                 t.Start(config);
             }
 
-            private void RunThread(object parameter)
+            private void RunDataServerThread(object parameter)
             {
                 DfsConfiguration config = (DfsConfiguration)parameter;
                 DataServer server = new DataServer(config);
@@ -79,15 +83,20 @@ namespace Tkl.Jumbo.Dfs.Test
             }
         }
 
-        public TestDfsCluster(int dataNodes, int replicationFactor)
-            : this(dataNodes, replicationFactor, null)
+        public TestDfsCluster(int dataServers, int replicationFactor)
+            : this(dataServers, replicationFactor, null)
         {
         }
 
-        public TestDfsCluster(int dataNodes, int replicationFactor, int? blockSize)
+        public TestDfsCluster(int dataServers, int replicationFactor, int? blockSize)
+            : this(dataServers, replicationFactor, blockSize, true)
+        {
+        }
+
+        public TestDfsCluster(int dataServers, int replicationFactor, int? blockSize, bool eraseExistingData)
         {
             string path = Utilities.TestOutputPath;
-            if( System.IO.Directory.Exists(path) )
+            if( eraseExistingData && System.IO.Directory.Exists(path) )
                 System.IO.Directory.Delete(path, true);
             System.IO.Directory.CreateDirectory(path);
 
@@ -97,20 +106,25 @@ namespace Tkl.Jumbo.Dfs.Test
             
             _clusterDomain = AppDomain.CreateDomain("TestCluster", null, setup);
 
-            _nameServerRunner = (NameServerRunner)_clusterDomain.CreateInstanceAndUnwrap(typeof(NameServerRunner).Assembly.FullName, typeof(NameServerRunner).FullName);
-            _nameServerRunner.Run(path, replicationFactor, dataNodes, blockSize);
+            _clusterRunner = (ClusterRunner)_clusterDomain.CreateInstanceAndUnwrap(typeof(ClusterRunner).Assembly.FullName, typeof(ClusterRunner).FullName);
+            _clusterRunner.Run(path, replicationFactor, dataServers, blockSize);
 
         }
 
         public void Shutdown()
         {
             Thread.Sleep(1000);
-            _nameServerRunner.Shutdown();
-            _nameServerRunner = null;
+            _clusterRunner.Shutdown();
+            _clusterRunner = null;
             Trace.WriteLine("Shutting down now.");
             Trace.Flush();
             AppDomain.Unload(_clusterDomain);
             _clusterDomain = null;
+        }
+
+        public void StartDataServers(int dataServers)
+        {
+            _clusterRunner.StartDataServers(dataServers);
         }
 
         public static DfsConfiguration CreateClientConfig()
